Split status logic into internal
