+++ internal/domain/domain_test.go
@@ -1,4 +1,4 @@
-package status
+package domain
 
 import (
 	"errors"
@@ -9,7 +9,6 @@ import (
 	"libvirt.org/go/libvirt"
 )
 
-// mockDomain implements Domain interface
 type mockDomain struct {
 	infoResult      *libvirt.DomainInfo
 	infoErr         error
@@ -21,20 +20,15 @@ type mockDomain struct {
 	guestInfoErr    error
 }
 
-func (m *mockDomain) GetInfo() (*libvirt.DomainInfo, error) {
-	return m.infoResult, m.infoErr
-}
+func (m *mockDomain) GetInfo() (*libvirt.DomainInfo, error) { return m.infoResult, m.infoErr }
 func (m *mockDomain) GetState() (libvirt.DomainState, int, error) {
 	return m.stateResult, 0, m.stateErr
 }
-func (m *mockDomain) GetUUID() ([]byte, error) {
-	return m.uuidResult, m.uuidErr
-}
+func (m *mockDomain) GetUUID() ([]byte, error) { return m.uuidResult, m.uuidErr }
 func (m *mockDomain) GetGuestInfo(types libvirt.DomainGuestInfoTypes, flags uint32) (*libvirt.DomainGuestInfo, error) {
 	return m.guestInfoResult, m.guestInfoErr
 }
 
-// mockConnect implements Connect interface
 type mockConnect struct {
 	domains []libvirt.Domain
 	err     error
@@ -44,10 +38,9 @@ func (m *mockConnect) ListAllDomains(flags libvirt.ConnectListAllDomainsFlags) (
 	return m.domains, m.err
 }
 
-// valid 16-byte UUID for testing
 var testUUIDBytes = []byte{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15}
 
-func TestDomainInfo_GetInfo_Success(t *testing.T) {
+func TestInfo_GetInfo_Success(t *testing.T) {
 	mock := &mockDomain{
 		infoResult: &libvirt.DomainInfo{
 			State:     1,
@@ -57,7 +50,7 @@ func TestDomainInfo_GetInfo_Success(t *testing.T) {
 			CpuTime:   100,
 		},
 	}
-	di := &DomainInfo{}
+	di := &Info{}
 	if err := di.GetInfo(mock); err != nil {
 		t.Fatalf("expected nil, got %v", err)
 	}
@@ -66,9 +59,9 @@ func TestDomainInfo_GetInfo_Success(t *testing.T) {
 	}
 }
 
-func TestDomainInfo_GetInfo_Error(t *testing.T) {
+func TestInfo_GetInfo_Error(t *testing.T) {
 	mock := &mockDomain{infoErr: fmt.Errorf("libvirt error")}
-	err := (&DomainInfo{}).GetInfo(mock)
+	err := (&Info{}).GetInfo(mock)
 	if err == nil {
 		t.Fatal("expected error, got nil")
 	}
@@ -77,13 +70,13 @@ func TestDomainInfo_GetInfo_Error(t *testing.T) {
 	}
 }
 
-func TestDomainState_GetInfo_Success(t *testing.T) {
+func TestState_GetInfo_Success(t *testing.T) {
 	mock := &mockDomain{
 		stateResult:     libvirt.DomainState(1),
 		uuidResult:      testUUIDBytes,
 		guestInfoResult: &libvirt.DomainGuestInfo{},
 	}
-	ds := &DomainState{}
+	ds := &State{}
 	if err := ds.GetInfo(mock); err != nil {
 		t.Fatalf("expected nil, got %v", err)
 	}
@@ -92,32 +85,32 @@ func TestDomainState_GetInfo_Success(t *testing.T) {
 	}
 }
 
-func TestDomainState_GetInfo_StateError(t *testing.T) {
+func TestState_GetInfo_StateError(t *testing.T) {
 	mock := &mockDomain{stateErr: fmt.Errorf("state error")}
-	err := (&DomainState{}).GetInfo(mock)
+	err := (&State{}).GetInfo(mock)
 	if !errors.Is(err, virerr.DomainStatusError) {
 		t.Errorf("expected DomainStatusError, got %v", err)
 	}
 }
 
-func TestDomainState_GetInfo_UUIDError(t *testing.T) {
+func TestState_GetInfo_UUIDError(t *testing.T) {
 	mock := &mockDomain{
 		stateResult: libvirt.DomainState(1),
 		uuidErr:     fmt.Errorf("uuid error"),
 	}
-	err := (&DomainState{}).GetInfo(mock)
+	err := (&State{}).GetInfo(mock)
 	if !errors.Is(err, virerr.InvalidUUID) {
 		t.Errorf("expected InvalidUUID, got %v", err)
 	}
 }
 
-func TestDomainState_GetInfo_GuestInfoError(t *testing.T) {
+func TestState_GetInfo_GuestInfoError(t *testing.T) {
 	mock := &mockDomain{
 		stateResult:  libvirt.DomainState(1),
 		uuidResult:   testUUIDBytes,
 		guestInfoErr: fmt.Errorf("guest info error"),
 	}
-	err := (&DomainState{}).GetInfo(mock)
+	err := (&State{}).GetInfo(mock)
 	if !errors.Is(err, virerr.DomainStatusError) {
 		t.Errorf("expected DomainStatusError, got %v", err)
 	}
@@ -143,7 +136,7 @@ func TestAllInstInfo_GetAllinstInfo_EmptyList(t *testing.T) {
 }
 
 func TestDataTypeRouter_ValidTypes(t *testing.T) {
-	cases := []DomainDataType{DomState, BasicInfo, GuestInfoUser, GuestInfoOS, GuestInfoFS, GuestInfoDisk}
+	cases := []DataType{DomState, BasicInfo, GuestInfoUser, GuestInfoOS, GuestInfoFS, GuestInfoDisk}
 	for _, c := range cases {
 		if _, err := DataTypeRouter(c); err != nil {
 			t.Errorf("DataTypeRouter(%d) returned error: %v", c, err)
@@ -152,7 +145,7 @@ func TestDataTypeRouter_ValidTypes(t *testing.T) {
 }
 
 func TestDataTypeRouter_InvalidType(t *testing.T) {
-	_, err := DataTypeRouter(DomainDataType(99))
+	_, err := DataTypeRouter(DataType(99))
 	if !errors.Is(err, virerr.InvalidParameter) {
 		t.Errorf("expected InvalidParameter, got %v", err)
 	}
