--- services/status/domain_status.go
@@ -1,75 +1,0 @@
-package status
-
-import (
-	"errors"
-	"fmt"
-
-	virerr "github.com/easy-cloud-Knet/KWS_Core/internal/error"
-	"github.com/google/uuid"
-	"libvirt.org/go/libvirt"
-)
-
-func (DI *DomainInfo) GetInfo(domain Domain) error {
-	info, err := domain.GetInfo()
-	if err != nil {
-		return virerr.ErrorGen(virerr.DomainStatusError, err)
-	}
-	DI.State = info.State
-	DI.MaxMem = info.MaxMem
-	DI.Memory = info.Memory
-	DI.NrVirtCpu = info.NrVirtCpu
-	DI.CpuTime = info.CpuTime
-	//basic info can be added
-	return nil
-}
-
-func (DP *DomainState) GetInfo(domain Domain) error {
-	info, _, err := domain.GetState()
-	//searching for coresponding second parameter, "Reason"
-	if err != nil {
-		return virerr.ErrorGen(virerr.DomainStatusError, err)
-	}
-
-	uuidBytes, err := domain.GetUUID()
-	if err != nil {
-		return virerr.ErrorGen(virerr.InvalidUUID, err)
-	}
-	uuidParsed, err := uuid.FromBytes(uuidBytes)
-	if err != nil {
-		return virerr.ErrorGen(virerr.InvalidUUID, err)
-	}
-
-	DP.DomainState = info
-	DP.UUID = string(uuidParsed.String())
-	userInfo, err := domain.GetGuestInfo(libvirt.DOMAIN_GUEST_INFO_USERS, 0)
-	if err != nil {
-		return virerr.ErrorGen(virerr.DomainStatusError, fmt.Errorf("error retreving guest info: %w", err))
-	}
-	DP.Users = userInfo.Users
-	return nil
-}
-
-func DomainDetailFactory(Handler DataTypeHandler, dom Domain) *DomainDetail {
-	return &DomainDetail{
-		DataHandle: Handler,
-		Domain:     dom,
-	}
-}
-
-func DataTypeRouter(types DomainDataType) (DataTypeHandler, error) {
-	switch types {
-	case DomState:
-		return &DomainState{}, nil
-	case BasicInfo:
-		return &DomainInfo{}, nil
-	case GuestInfoUser:
-		return &DomainInfo{}, nil
-	case GuestInfoOS:
-		return &DomainInfo{}, nil
-	case GuestInfoFS:
-		return &DomainInfo{}, nil
-	case GuestInfoDisk:
-		return &DomainInfo{}, nil
-	}
-	return nil, virerr.ErrorGen(virerr.InvalidParameter, errors.New("invalid flag for DataRoute entereed "))
-}
